Rename hpxcube module to hpxnd.
