+++ gammapy/maps/hpxnd.py
